Move database config to dedicated configuration class
All other configurations have been moved to dedicated classes, so it
makes sense to so this for the database config, too.
This is also a preparation for the upcoming CLI.
