+++ src/main/java/org/osiam/configuration/DatabaseConfiguration.java
@@ -1,0 +1,56 @@
+package org.osiam.configuration;
+
+import com.zaxxer.hikari.HikariConfig;
+import com.zaxxer.hikari.HikariDataSource;
+import org.flywaydb.core.Flyway;
+import org.flywaydb.core.api.MigrationVersion;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+
+import javax.sql.DataSource;
+
+@Configuration
+@EnableTransactionManagement
+public class DatabaseConfiguration {
+
+    @Value("${org.osiam.db.driver}")
+    private String driverClassName;
+
+    @Value("${org.osiam.db.url}")
+    private String databaseUrl;
+
+    @Value("${org.osiam.db.username}")
+    private String databaseUserName;
+
+    @Value("${org.osiam.db.password}")
+    private String databasePassword;
+
+    @Value("${org.osiam.db.vendor}")
+    private String databaseVendor;
+
+    @Primary
+    @Bean
+    public DataSource dataSource() {
+        HikariConfig hikariConfig = new HikariConfig();
+        hikariConfig.setPoolName("osiam-cp");
+        hikariConfig.setDriverClassName(driverClassName);
+        hikariConfig.setJdbcUrl(databaseUrl);
+        hikariConfig.setUsername(databaseUserName);
+        hikariConfig.setPassword(databasePassword);
+        return new HikariDataSource(hikariConfig);
+    }
+
+    @Bean
+    public Flyway flyway() {
+        Flyway flyway = new Flyway();
+        flyway.setDataSource(dataSource());
+        flyway.setLocations("db/migration/" + databaseVendor);
+        flyway.setBaselineOnMigrate(true);
+        flyway.setBaselineVersion(MigrationVersion.fromVersion("1"));
+        flyway.migrate();
+        return flyway;
+    }
+}
